requirements.txt
flask==2.0.1
flask-pymongo==2.3.0
python-dotenv==0.19.0
pymongo[srv]
werkzeug==2.0.1